pyproject.toml
[tool.poetry]
name = "zmq-demo"
version = "0.1.0"
description = ""
authors = ["Soof Golan <hi@soofgolan.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyzmq = "^25.1.1"
typer = {extras = ["all"], version = "^0.9.0"}
fastapi = "^0.103.1"
tqdm = "^4.66.1"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
